+++ storageBuilding.go
@@ -7,6 +7,6 @@ type storageBuilding struct {
 }
 
 // Capacity returns the capacity of a storage building
-func (s storageBuilding) Capacity(lvl int) int {
-	return 5000 * int(2.5*math.Pow(math.E, (20*float64(lvl))/33))
+func (s storageBuilding) Capacity(lvl int64) int64 {
+	return 5000 * int64(2.5*math.Pow(math.E, (20*float64(lvl))/33))
 }
